add (and default to on) --preview
With this flag, the benchdiff output is rendered between each
interleaved run of both the old and new benches (within the current
package).

Especially for longer benchmark runs, this is helpful to give an idea
of what the results are likely going to be. In some cases, this may
prompt the user to call it early, thus saving time. But it's also
pleasant to be able to just check in, which is my primary motivation
for this change.