Merge profiles
Previously, `--{mem,cpu,mutex}profile` would instruct each iteration
of the benchmark to write the profile to the same file, leaving only
the last iteration's output in place.

This commit instead clears all output profiles before the first run, and then
merges all consecutive profiles into each other. This means that it is now
meaningful to combine `--count` and `--{mem,cpu,mutex}profile`; one will get
a result and a set of profiles that exactly corresponds to the result.

The previous workflow was to run an extra invocation with `--count=1` at the
end to collect profiles, but then the new/old profiles would not compare as
well since they hadn't been smoothed out over multiple runs.

Now, when profiling flags are passed, differences and profiles correspond
to each other, which I've wished for several times in the past.